+++ Assets/PatchKit Patcher/Scripts/AppData/Remote/IRemoteData.cs
@@ -7,14 +7,15 @@
         /// </summary>
         /// <param name="versionId">The version id.</param>
         /// <param name="keySecret">The key secret.</param>
-        RemoteResource GetContentPackageResource(int versionId, string keySecret);
+        /// <param name="countryCode"></param>
+        RemoteResource GetContentPackageResource(int versionId, string keySecret, string countryCode);
 
         /// <summary>
         /// Returns the diff package resource.
         /// </summary>
         /// <param name="versionId">The version id.</param>
         /// <param name="keySecret">The key secret.</param>
-        RemoteResource GetDiffPackageResource(int versionId, string keySecret);
+        RemoteResource GetDiffPackageResource(int versionId, string keySecret, string countryCode);
 
         /// <summary>
         /// Returns password of the content package resource.