+++ Assets/PatchKit Patcher/Scripts/AppData/Remote/RemoteData.cs
@@ -29,7 +29,7 @@ namespace PatchKit.Unity.Patcher.AppData.Remote
             _mainApiConnection = mainApiConnection;
         }
 
-        public RemoteResource GetContentPackageResource(int versionId, string keySecret)
+        public RemoteResource GetContentPackageResource(int versionId, string keySecret, string countryCode)
         {
             Checks.ArgumentValidVersionId(versionId, "versionId");
 
@@ -41,7 +41,7 @@ namespace PatchKit.Unity.Patcher.AppData.Remote
 
             var summary = _mainApiConnection.GetAppVersionContentSummary(_appSecret, versionId);
             var torrentUrl = _mainApiConnection.GetAppVersionContentTorrentUrl(_appSecret, versionId, keySecret);
-            var urls = _mainApiConnection.GetAppVersionContentUrls(_appSecret, versionId); // TODO: Add key secret checking
+            var urls = _mainApiConnection.GetAppVersionContentUrls(_appSecret, versionId, countryCode); // TODO: Add key secret checking
 
             resource.Size = summary.Size;
             resource.HashCode = summary.HashCode;
@@ -53,7 +53,7 @@ namespace PatchKit.Unity.Patcher.AppData.Remote
             return resource;
         }
 
-        public RemoteResource GetDiffPackageResource(int versionId, string keySecret)
+        public RemoteResource GetDiffPackageResource(int versionId, string keySecret, string countryCode)
         {
             Checks.ArgumentValidVersionId(versionId, "versionId");
 
@@ -65,7 +65,7 @@ namespace PatchKit.Unity.Patcher.AppData.Remote
 
             var summary = _mainApiConnection.GetAppVersionDiffSummary(_appSecret, versionId);
             var torrentUrl = _mainApiConnection.GetAppVersionDiffTorrentUrl(_appSecret, versionId, keySecret);
-            var urls = _mainApiConnection.GetAppVersionDiffUrls(_appSecret, versionId); // TODO: Add key secret checking
+            var urls = _mainApiConnection.GetAppVersionDiffUrls(_appSecret, versionId, countryCode); // TODO: Add key secret checking
 
             resource.Size = summary.Size;
             resource.HashCode = summary.HashCode;