+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Using geolocation to find closest http server available
+- Decides on torrents based on publish_method remote app property
+
 ### Fixed
 - HTTP downloader resumes the download on error
 - Moving files after installation instead of copying them